% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\docType{data}
\name{allowed_whisper_models}
\alias{allowed_whisper_models}
\title{Allowed Whisper models for transcription.}
\format{
An object of class \code{character} of length 6.
}
\usage{
allowed_whisper_models
}
\description{
Allowed Whisper models for transcription.
}
\keyword{datasets}
